move empty-leaf rewrite from broker into PinotDispatchPlanner
